Crate fixedbitset

FixedBitSet is a simple fixed size set of bits.

Crate features

std (default feature)
Disabling this feature disables using std and instead uses crate alloc. Requires Rust 1.36 to disable.

Rust Version

This version of fixedbitset requires Rust 1.39 or later.

Structs

Difference

An iterator producing elements in the difference of two sets.

FixedBitSet

FixedBitSet is a simple fixed size set of bits that each can be enabled (1 / true) or disabled (0 / false).

Intersection

An iterator producing elements in the intersection of two sets.

Ones

An iterator producing the indices of the set bit in a set.

SymmetricDifference

An iterator producing elements in the symmetric difference of two sets.

Union

An iterator producing elements in the union of two sets.

Traits

IndexRange

IndexRange is implemented by Rust’s built-in range types, produced by range syntax like .., a.., ..b or c..d.