Crate fixedbitset

FixedBitSet is a simple fixed size set of bits.

Modules

range 🔒

Structs

Difference

An iterator producing elements in the difference of two sets.

FixedBitSet

FixedBitSet is a simple fixed size set of bits that each can be enabled (1 / true) or disabled (0 / false).

Intersection

An iterator producing elements in the intersection of two sets.

Masks 🔒

Ones

An iterator producing the indices of the set bit in a set.

Union

An iterator producing elements in the union of two sets.

Constants

BITS 🔒

Statics

FALSE 🔒

TRUE 🔒

Traits

IndexRange

IndexRange is implemented by Rust’s built-in range types, produced by range syntax like .., a.., ..b or c..d.

Functions

div_rem 🔒

Type Aliases

Block 🔒