Struct fixedbitset::FixedBitSet

pub struct FixedBitSet { /* fields omitted */ }

FixedBitSet is a simple fixed size set of bits that can be enabled (1 / true) or disabled (0 / false).

Methods

impl FixedBitSet

fn with_capacity(bits: usize) -> Self

Create a new FixedBitSet with a specific number of bits, all initially clear.

fn grow(&mut self, bits: usize)

Grow capacity to bits, all new bits initialized to zero

fn len(&self) -> usize

Return the length of the FixedBitSet in bits.

fn contains(&self, bit: usize) -> bool

Return true if the bit is enabled in the FixedBitSet, false otherwise.

Note: bits outside the capacity are always disabled.

Note: Also available with index syntax: bitset[bit].

fn clear(&mut self)

Clear all bits.

fn insert(&mut self, bit: usize)

Enable bit.

Panics if bit is out of bounds.

fn put(&mut self, bit: usize) -> bool

Enable bit, and return its previous value.

Panics if bit is out of bounds.

fn set(&mut self, bit: usize, enabled: bool)

Panics if bit is out of bounds.

fn copy_bit(&mut self, from: usize, to: usize)

Copies boolean value from specified bit to the specified bit.

Panics if to is out of bounds.

fn count_ones<T: IndexRange>(&self, range: T) -> usize

Count the number of set bits in the given bit range.

Use .. to count the whole content of the bitset.

Panics if the range extends past the end of the bitset.

fn as_slice(&self) -> &[u32]

View the bitset as a slice of u32 blocks

fn as_mut_slice(&mut self) -> &mut [u32]

View the bitset as a mutable slice of u32 blocks. Writing past the bitlength in the last will cause contains to return potentially incorrect results for bits past the bitlength.