Struct fixedbitset::FixedBitSet

pub struct FixedBitSet { /* fields omitted */ }

FixedBitSet is a simple fixed size set of bits that each can be enabled (1 / true) or disabled (0 / false).

The bit set has a fixed capacity in terms of enabling bits (and the capacity can grow using the grow method).

Methods

impl FixedBitSet

pub fn with_capacity(bits: usize) -> Self

Create a new FixedBitSet with a specific number of bits, all initially clear.

pub fn grow(&mut self, bits: usize)

Grow capacity to bits, all new bits initialized to zero

pub fn len(&self) -> usize

Return the length of the FixedBitSet in bits.

pub fn contains(&self, bit: usize) -> bool

Return true if the bit is enabled in the FixedBitSet, false otherwise.

Note: bits outside the capacity are always disabled.

Note: Also available with index syntax: bitset[bit].

pub fn clear(&mut self)

Clear all bits.

pub fn insert(&mut self, bit: usize)

Enable bit.

Panics if bit is out of bounds.

pub fn put(&mut self, bit: usize) -> bool

Enable bit, and return its previous value.

Panics if bit is out of bounds.

pub fn set(&mut self, bit: usize, enabled: bool)

Panics if bit is out of bounds.

pub fn copy_bit(&mut self, from: usize, to: usize)

Copies boolean value from specified bit to the specified bit.

Panics if to is out of bounds.

pub fn count_ones<T: IndexRange>(&self, range: T) -> usize

Count the number of set bits in the given bit range.

Use .. to count the whole content of the bitset.

Panics if the range extends past the end of the bitset.

pub fn set_range<T: IndexRange>(&mut self, range: T, enabled: bool)

Sets every bit in the given range to the given state (enabled)

Use .. to toggle the whole bitset.

Panics if the range extends past the end of the bitset.

pub fn insert_range<T: IndexRange>(&mut self, range: T)

Enables every bit in the given range.

Use .. to make the whole bitset ones.

Panics if the range extends past the end of the bitset.

pub fn as_slice(&self) -> &[u32]

View the bitset as a slice of u32 blocks

pub fn as_mut_slice(&mut self) -> &mut [u32]

View the bitset as a mutable slice of u32 blocks. Writing past the bitlength in the last will cause contains to return potentially incorrect results for bits past the bitlength.

Important traits for Ones<'a>
impl<'a> Iterator for Ones<'a> type Item = usize;

pub fn ones(&self) -> Ones

Iterates over all enabled bits.

Iterator element is the index of the 1 bit, type usize.

Important traits for Intersection<'a>
impl<'a> Iterator for Intersection<'a> type Item = usize;

pub fn intersection<'a>(&'a self, other: &'a FixedBitSet) -> Intersection<'a>

Returns a lazy iterator over the intersection of two FixedBitSets

Important traits for Union<'a>
impl<'a> Iterator for Union<'a> type Item = usize;

pub fn union<'a>(&'a self, other: &'a FixedBitSet) -> Union<'a>

Returns a lazy iterator over the union of two FixedBitSets.

Important traits for Difference<'a>
impl<'a> Iterator for Difference<'a> type Item = usize;

pub fn difference<'a>(&'a self, other: &'a FixedBitSet) -> Difference<'a>

Returns a lazy iterator over the difference of two FixedBitSets. The difference of a and b is the elements of a which are not in b.