Struct fixedbitset::FixedBitSet

pub struct FixedBitSet {
    // some fields omitted
}

Methods

impl FixedBitSet

fn with_capacity(bits: usize) -> FixedBitSet

Create a new FixedBitSet with a specific number of bits, all initially clear.

fn len(&self) -> usize

Return the length of the FixedBitSet.

fn contains(&self, bit: usize) -> bool

Return true if the bit is enabled in the FixedBitSet, false otherwise.

Note: bits outside the capacity are always disabled.

fn clear(&mut self)

fn insert(&mut self, bit: usize)

Panics if bit is out of bounds.

fn set(&mut self, bit: usize, enabled: bool)

Panics if bit is out of bounds.